Add files via upload
This python script allows you to launch an LFI attack against any path that you suspect vulnerable. To use this script, you need just to change the payload file and add your own payload file and change also the the vulnerable path to add the path you suspect vulnerable and finally launch the script . This a very simple useful script that can help a pentester to automate LFI exploit vulnerabilities. Feel free to download, modify it or improve it. Thanks.